xxhash: rename XXH_USELESS_ALIGN_BRANCH to XXH_FORCE_ALIGN_CHECK
The name change also includes inverting the logic, which makes the evaluation
logic clearer: if FORCE_ALIGN_CHECK is set, we perform the check;
if not, we don't.

While at it, allow calling code to set the flag. The use case for this
is the following: calling code might have a guarantee that the input data
is aligned. Thus, for maximum performance, it would set:

  XXH_FORCE_ALIGN_CHECK to 0, to skip the alignment check
  XXH_FORCE_MEMORY_ACCESS 2, to copy data in chunks of 32 or 64 bits

Without this change, there is no way to skip the alignment check on
certain architectures.

Signed-off-by: Emilio G. Cota <cota@braap.org>