Fixes: NEON intrinsic, typo fix, revert XXH64, use non-vectorized on unaligned x86
On NEON, both compilers fail to generate vshr.u32/vsli.32 in XXH_rotlvec32, and
will instead emit vshr.u32/vshl.u32/vorr, which is slower. Now, I use an
intrinsic to force a vsli.32 instruction.

I think it is acceptable if there are one or two intrinsics if it is beneficial
for performance. I just don't want this, as it hampers readability significantly.

    #if defined(__SSE4_1__)
        __m128i inp = _mm_loadu_si128((const __m128i*)p);
        ...
    #elif defined(__ARM_NEON__) || defined(__ARM_NEON)
        uint32x4_t inp = vld1q_u32((const U32*)p);
        ...

I had quite a few dumb typos, namely this:

    return (x << right) | (x >> left);

and this:

    U32x4 prime1 = { PRIME32_2, PRIME32_1, PRIME32_1, PRIME32_1 };

That is what I get for coding late at night. *facepalms*

XXH64 vectorization was pointless. The reason it was so slow was because
SSE4.1 lacks a U64x2 'type' completely. It doesn't matter how much you
optimize it, emulating int math is always going to be slower than hardware.

On x86, unaligned SSE reads are much slower than the aligned SSE reads,
and the non-vectorized one is slightly faster. (2.6 vs 2.9 GB/s)

However, on NEON, the old version was very slow and we use vectorized
instructions regardless.