Separate into XXH32a, add xxhsum tests, fix some ARM and Thumb weirdness.
XXH32 is restored, the new vectorized XXH32 is called XXH32a.
The rest of XXH32a is pretty much done.

XXH32a will fall back to the old style when endianness does not match,
and on x86_64, when things aren't aligned.

xxhsum now has the proper tests and hash values.

I fixed a typo in the ARM intrinsics.

The main loops now check endianness and will change to a basic dereference
instead of XXH_readxxx when the pointer is aligned and endianness matches.
This prevents branching on Thumb.

GCC will be stupid when building for Thumb-1 and convert every constant
multiplication into shift/add, even if it is significantly slower than
just doing the multiplication normally. We force GCC to see the primes
as mutable, so it can't perform this "optimization". This makes Thumb-1
code 5-6 times faster.