Add XXH64a, improve performance
XXH32a and XXH64a use the same internals and hash functions, but
the beginning and end are different.

This means that XXH64a is going to be just as fast as XXH32a on
32-bit machines, and still have good performance on 64-bit machines.

I also used an ugly __asm__ hack to prevent unwanted vectorization
in XXH32, without affecting the rest of the program or the previous
performance. This was the only thing that worked so far.

I also made some changes to improve performance, such as reordering
things and changing XXH_IS_LITTLE_ENDIAN to use __LITTLE_ENDIAN__
or __BIG_ENDIAN__ first instead of doing runtime checks.