xxh3: modified 9-16 bytes formula
ok, let's bite the bullet.
I believe it's not useful to attempt to block a multiplication by zero for "security" purpose.
The only useful impact is to tame the space reduction.

On the 9-16 range (for the _64bits variant), space reduction is very minimal.
It's trival to show that all 64-bit values can be generated (consider a 1x multiplier),
so there is not a single miss in the spectre.

What is uneven is that we don't have a homogeneous 1/2^64 distribution over all values.
But even then, with the multiplication by zero let loose,
the worst offender would be the value Avalanche(len + zero), and its share is 2/2^64.
Aka, barely above the average. Everything else will be close to 1.
Even if we end up with 63.9 bits of entropy instead of 64, I could not care less.

We have to keep in mind the point of a hash algorithm for small inputs :
it's all about generating a bit field, to index a table, or bloom filters.
Which means, the entire 64-bits value is not even used "in full" (in contrast with checksumming).
A few bits (typically 10-20) will be extracted from it. Eventually several ones.
At this scale, the distribution already looks perfect.
I don't think we are improving anything substantial by neuteuring the multiplication by zero,
but for the sake of math, and since the measured cost is so small, okay, let's do it.