[PPC64][WIP] Fix it, make it compatible with POWER8, and don't allow POWER7.
The VSX codepath now works fine, and seems to compile on GCC 4.8.5.

The POWER9 code is now optional, and the only time it is used is for
swapping on big endian.

On big endian, we now use vperm and vpermxor to combine the swap with
the data_key xor.

About to start testing on actual POWER8 hardware, to make sure everything
works.