Document main loop, place most data/key with input/secret, merge SSE2/AVX2/sync with NEON
With the exception of the inner accumulate loops (where it becomes
ugly to describe), data and key have been almost completely replaced
with input and secret.

SSE2 and AVX2 now use macros to reuse the same implementation:

    XXH_vec_t x = XXH_loadu_si(ptr);
    x = XXH_add_epi64(x, y);
    x = XXH_shuffle_epi32(x, _MM_SHUFFLE(2,3,0,1));
    ...

This means fewer algorithms to break.

They have also been reordered and redocumented to be consistent with the
NEON version. I want VSX to look like that as well.

A workaround for -Wcast-align and _mm_loadu_si128 was inserted, as much
as I dislike it.