Abstract xxhsum bench, increase line length
Uses a table and a loop to reduce copy/paste and allow easy testing of
other hash functions. Create a wrapper and insert it into
g_hashesToBench, and it will automatically be added to the benchmark.

The hash display line has been made longer to actually fit xxHash's
names instead of clipping them. It is also configurable.