remove alignment check on aarch64
as this architecture offers decent performance for unaligned memory accesses (like x86/x64).

This circumvent the problem described in #383,
as the direct read path using `const xxh_u32*` will not be generated by default.

Note however that this does not go at the root of the problem,
which is a possibility of strict-aliasing issue
when using the direct read path _and_ inlining xxh32.
This issue however is more difficult to solve without a reproduction case.

I also used this opportunity to fix or reinforce documentation.