xxhsum: Add option for disabling benchmarks
Similar to XSUM_NO_TESTS, if the benchmark functions are called when
compiled with this option, it will print an error and exit 1.

If xxhsum.c was compiled with XSUM_NO_TESTS, the benchmark options will
be removed from the usage. However, only xsum_bench.c needs to be
compiled with this option.