added variant XXH3_64bits_withSecretandSeed()
and its companion XXH3_generateSecret_fromSeed().

The new variant uses @seed for "small" keys,
and @secret for "large" keys.
When combined with the new generator,
it provides exactly the same results as _withSeed(),
with the benefit of pre-caculated @secret for "large keys",
resulting in a speed boost for "not too large" keys (<1 KB).