[NEON] Enable XXH_NEON for MSVC ARM/ARM64
When MSVC is targeting ARMv7VE or ARM64, it will now use the NEON path.

It actually just works(TM) out of the box.

The code hasn't been tested, as I don't have access to an ARM64 Windows
device, but the code appears to be correct judging from the assembly
listings.