fix #625
PVS Studio static analyzer doesn't like that the structure is not entirely initialized.
This was supposed to allow, in some future, to reduce the size of the state,
while minimizing impact with library's installed based (in case of version mismatch).

This is way too complex and risky for such a tiny benefit.
So revert to the classical "init everything" method, which should please PVS Studio.
The last member will remain "reserved" and unused.
Who knows, maybe in some future, it may be of some use, for some purpose yet to determine.