Use `vpsrlq` instead of `vpshufd` in avx2/avx512 accumulate.
`vpshufd` was just to essentially implement `dst >> 32` which can (and
should) be done with `vpsrlq`.

`XXH3_{32|64|128}bits` throughput seems to be limitted by a port_5
bottleneck and under-uses port_0.

Hashing 1mb in a loop 100000 times with `vpshufd`:
    x86-64-v4:
         5,363,095,005      cycles
         5,195,740,817      port_0
            17,104,516      port_1
         5,358,811,641      port_5
    x86-64-v3:
         7,954,143,671      cycles
         6,610,562,251      port_0
         7,405,810,291      port_1
         7,474,008,658      port_5

`vpshufd` on x86-64-v3 only runs on port_5 (and for x86-64-v4
only runs on port_5 when used with `zmm` registers). `vpsrlq` can
run on port_{0|1} (port_0 only for `zmm`) which helps to
alleviate this bottleneck.

Hashing 1mb in a loop 100000 times with `vpsrlq`:
    x86-64-v4:
         5,305,723,460      cycles
         5,246,052,997      port_0
            17,759,291      port_1
         5,303,090,711      port_5
    x86-64-v3:
         7,852,139,442      cycles
         6,950,368,526      port_0
         7,301,134,590      port_1
         7,215,987,871      port_5

which in general gets a better port distribution and results in
slightly better times.

Note the sse2 accumulate also would benefit from shifting uops from
port5 -> port0

    x86-64-v2 with `pshufd`:
        13,128,826,786      port_0
        15,014,228,537      port_1
        14,801,410,216      port_5

    x86-64-v2 with `psrlq`:
        13,899,468,407      port_0
        14,774,790,828      port_1
        14,555,300,082      port_5

But since `pshufd` has a 3-term encoding and `psrlq` only has a 2-term
encoding, using `psrlq` causes an extra `movdqa` in codegen which
is a net-negative (at least on Icelake benchmarks).

Benchmarking on Icelake this commit results in the following
improvements:

NB: The avx2 changes where also benchmarked on Icelake which is
not really ideal. Although note that Icelake (unlike native
x86-64-v3 hardware has a faster `vpshufd` (port_{1|5} vs
just port_5) so hopefully anything skewed by incorrect hardware
was infavor of the original implementation).