Fix `XXH_UNREACHABLE` feature test for C++23 and remove for C23 #791
gcc11/12 don't support `std::unreachable()` in there `-std=c++23`
build.

A more reliable check is:
```
```
so use that.

A similiar issue may occur for C23 `unreachable()` so just remove it
on the basis that if a compiler doesn't currently support some
builtin for unreachable/assume its very likely not using it in
analysis so we are probably not missing much (if anything)
leaving it out.