[NEON/VSX] Fix GCC strict aliasing issues
This fixes GCC 12 on macOS ARM64, as well as the existing s390x zVector
bug that was worked around with a `__sync_synchronize()` by using
`__attribute__((may_alias))` which excludes a type from strict aliasing
optimizations.

`vldq`/`vst1q` are still avoided due to older versions having issues with
load-store optimizations.

This is not needed on x86 because the SSE types are impossible to not
alias.