disable XXH64 autovectorization
using a compiler fence (like XXH32).

XXH64 has been reported to be auto-vectorized on AVX512 architecture.
Like XXH32, it's generally a bad idea to vectorize XXH64.

However, in order to test, or if the result differ for newer cpus,
it's possible to disable the fence by using XXH_ENABLE_AUTOVECTORIZE build variable.